Worked on charge issues with CD music. Now uses only charge->charge_balance for storing charge balance information.
Use of comp->cb is removed with SKIP_MUSIC ifdef.

Fixed a segv if element was not in database.

Uses master species for DDL calculation, expecting it to be charge
balanced. Actually tests masters species for this case and reports an
error if not.

Added line to give total equivalents of surface charge, sum of 0, 1,
and 2 planes.



git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk@1129 1feff8c3-07ed-0310-ac33-dd36852eb9cd