Worked some with solid solutions for small total of component. Tries to leave 1e-25 (MIN_TOTAL) and not crash. Uses MIN_TOTAL_SS, but currently equal to MIN_TOTAL.
Fixed the logic of rewriting phase equations containing solid and
gases. Needed trxn_add_phase, fix to trxn_copy, and force .s to 
be NULL for (g) or (s).

Changed print for solid solution, pure phases, and distribution of 
species to use longer fields for names.

Reran and checked test problems. S_S may run a little worse, but
results seem to be the same.






git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk/src@1306 1feff8c3-07ed-0310-ac33-dd36852eb9cd