Added -cell to dump, copy, delete.
Selects anything with the specified user numbers.

Fixed bug in read_class where end of file caused the
last line of input to be repeated in the output.

removed warning print from copy_entities, which is nicer 
for copy -cell and some reactants are not present.

rewrote read_copy to include -cell, but also changed logic
for selection of appropriate keywords. Now has list that are acceptable, whereas before it was a list of unacceptable.

git-svn-id: svn://136.177.114.72/svn_GW/phreeqc/trunk@4823 1feff8c3-07ed-0310-ac33-dd36852eb9cd